32bit compat

+++ pandas/tests/dtypes/cast/test_construct_ndarray.py
@@ -23,7 +23,7 @@ def test_construct_1d_ndarray_preserving_na(values, dtype, expected):
 
 @pytest.mark.parametrize("dtype", ["m8[ns]", "M8[ns]"])
 def test_construct_1d_ndarray_preserving_na_datetimelike(dtype):
-    arr = np.arange(5).view(dtype)
+    arr = np.arange(5, dtype=np.int64).view(dtype)
     expected = np.array(list(arr), dtype=object)
     assert all(isinstance(x, type(arr[0])) for x in expected)
